Refactor: dto 관련 패키지 생성하여 패키징
기존의 class 파일로 관리하던 파일을 record로 변경하였습니다.
