+++ src/main/java/roomescape/member/dto/AuthUserNameResponse.java
@@ -1,4 +1,4 @@
-package roomescape.member;
+package roomescape.member.dto;
 
 public record AuthUserNameResponse(String name) {
 }
