+++ src/main/java/roomescape/member/dto/LoginRequest.java
@@ -1,4 +1,4 @@
-package roomescape.member;
+package roomescape.member.dto;
 
 public record LoginRequest(
         String email,
