+++ src/main/java/roomescape/member/dto/LoginResponse.java
@@ -1,4 +1,4 @@
-package roomescape.member;
+package roomescape.member.dto;
 
 public record LoginResponse(String token) {
 }
