+++ src/main/java/roomescape/member/dto/MemberRequest.java
@@ -1,0 +1,4 @@
+package roomescape.member.dto;
+
+public record MemberRequest(String name, String email, String password) {
+}
