+++ src/main/java/roomescape/member/dto/MemberResponse.java
@@ -1,0 +1,4 @@
+package roomescape.member.dto;
+
+public record MemberResponse(Long id, String name, String email) {
+}
